Move the filter help from right panel to the dialog
In right panel, there was not enough space to display the help
correctly. So it was not useful because it was very hard to read.

Now, the help is displayed in the dialog in a <details> section.

When the filter dialog is open, it now displays the Workspace section in
right panel as it is more related.